requirements.txt
asgiref==3.7.2
Django==3.2
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.12.2
Faker==4.14.2
psycopg2-binary==2.9.6
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing_extensions==4.6.3
validate-docbr==1.8.2
